+++ src/debug-rendering/renderables/Paths.h
@@ -5,6 +5,8 @@ namespace devtools::debug_rendering::renderables {
 	class Paths : public OptionalDebugRenderable {
 	public:
 		bool enabled{};
+		bool normalsEnabled{};
+		bool tangentsEnabled{};
 
 		virtual void RenderIngameDebugVisuals(hh::gfnd::DrawContext& ctx) override;
 		virtual void RenderImGuiDebugVisuals(const ImGuiDrawContext& ctx) override;