+++ src/ui/Action.h
@@ -32,6 +32,7 @@ enum class ActionId {
 	LOCK_Y_AXIS,
 	ONLY_Z_AXIS,
 	LOCK_Z_AXIS,
+	SCENE_CHANGED,
 	DELETE,
 	OBJECT_PLACED,
 	TOGGLE_PLACE_MODE,
@@ -40,6 +41,10 @@ enum class ActionId {
 	LEVEL_EDITOR_STOP_CHANGING_PARAMETERS,
 	LEVEL_EDITOR_SET_FOCUSED_CHUNK,
 	LEVEL_EDITOR_FOCUSED_CHUNK_CHANGED,
+	LEVEL_EDITOR_PATHS_UPDATED,
+	LEVEL_EDITOR_OPEN_ARRAY_TOOL,
+	LEVEL_EDITOR_SET_OBJECT_CLASS_TO_PLACE,
+	LEVEL_EDITOR_OBJECT_CLASS_TO_PLACE_CHANGED,
 	FOCUS_GAME_OBJECT,
 	FOCUS_OBJECT_DATA,
 	EXPORT_RESOURCE,
@@ -65,9 +70,9 @@ template<typename A>
 struct AsyncAction : AsyncActionBase {
 	A action;
 
-	AsyncAction(csl::fnd::IAllocator* allocator, const A& action) : AsyncActionBase{ allocator }, action{ action.payload } {}
+	AsyncAction(csl::fnd::IAllocator* allocator, const A& action) : AsyncActionBase{ allocator }, action{ action } {}
 
-	//AsyncAction(csl::fnd::IAllocator* allocator, A&& action) : AsyncActionBase{ allocator }, action{ std::move(action.payload) } {}
+	//AsyncAction(csl::fnd::IAllocator* allocator, A&& action) : AsyncActionBase{ allocator }, action{ std::move(action) } {}
 
 	virtual ActionBase& GetAction() override { return action; }
 };
@@ -95,6 +100,8 @@ struct Action : ActionBase {
 	explicit Action(const std::remove_reference_t<Payload>& payload) : ActionBase{ actionId }, payload{ payload } {}
 	explicit Action(std::remove_reference_t<Payload>& payload) : ActionBase{ actionId }, payload{ payload } {}
 	explicit Action(std::remove_reference_t<Payload>&& payload) : ActionBase{ actionId }, payload{ std::move(payload) } {}
+	Action(const Action<actionId, Payload>& other) : ActionBase{ other.id }, payload{ other.payload } {}
+	Action(Action<actionId, Payload>&& other) : ActionBase{ other.id }, payload{ std::move(other.payload) } {}
 
 	virtual AsyncActionBase* CreateAsync(csl::fnd::IAllocator* allocator) const & override {
 		return new (allocator) AsyncAction<Action<actionId, Payload>>{ allocator, *this };
@@ -115,6 +122,8 @@ public:
 	virtual void ProcessAction(const ActionBase& action);
 };
 
+// This should become a SceneEditing behavior and Place/Delete should use it to edit the scene.
+using SceneChangedAction = Action<ActionId::SCENE_CHANGED>;
 using FocusGameObjectAction = Action<ActionId::FOCUS_GAME_OBJECT, hh::game::GameObject*>;
 using FocusObjectDataAction = Action<ActionId::FOCUS_OBJECT_DATA, hh::game::ObjectData*>;
 using ExportResourceAction = Action<ActionId::EXPORT_RESOURCE>;