+++ src/ui/common/editors/HsmBase.h
@@ -1,0 +1,3 @@
+#pragma once
+
+bool Editor(const char* label, hh::ut::HsmBase& hsm);