+++ src/ui/operation-modes/OperationModeHost.h
@@ -7,6 +7,8 @@ public:
 	virtual void EndSceneWindow() = 0;
 	virtual bool BeginOverlayWindow() = 0;
 	virtual void EndOverlayWindow() = 0;
+	virtual bool BeginMenuWindow() = 0;
+	virtual void EndMenuWindow() = 0;
 	virtual bool IsMouseOverSceneWindow() = 0;
 	virtual void Dispatch(const ActionBase& action) = 0;
 };