+++ src/ui/operation-modes/behaviors/Delete.h
@@ -39,6 +39,7 @@ public:
 
 			selectionBehavior->DeselectAll();
 			traits.DeleteObjects(tmpArr);
+			Dispatch(SceneChangedAction{});
 		}
 		}
 	}